+++ backend/ItemsService/ItemsService.csproj
@@ -24,4 +24,8 @@
     <ProjectReference Include="..\packages\TrendingApp.Packages.MassTransitDependencyInjection\TrendingApp.Packages.MassTransitDependencyInjection.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Context\" />
+  </ItemGroup>
+
 </Project>
